scripts/requirements.txt
requests
openai
python-dotenv
pdfplumber
flask
